Move main.go entrypoint to the cmd folder
